refactor(youtube): move youtube client to config
The youtube api client can now be set in configuration, this will allow
us to use mock versions in test.

Two implementations have been added, one "real" http client and another
that returns dummy data for tests.

See: http://blog.plataformatec.com.br/2015/10/mocks-and-explicit-contracts/
