+++ web/controllers/api_controller.ex
@@ -1,10 +1,11 @@
 defmodule Juicebox.ApiController do
   use Juicebox.Web, :controller
-
   import Juicebox.Youtube
 
+  @youtube_api Application.get_env(:juicebox, :youtube_api)
+
   def videos(conn, params) do
-    case Juicebox.Youtube.search(params["q"]) do
+    case @youtube_api.search(params["q"]) do
       {:ok, videos} ->
         render(conn, videos: videos)
       {:error, _} ->
